feat: add ARK Survival Ascended support
