+++ src/clients/ark-survival-ascended.client.ts
@@ -1,0 +1,3 @@
+import { ArkClient } from "./ark.client";
+
+export class ArkSurvivalAscendedClient extends ArkClient {}
