+++ src/game.ts
@@ -1,6 +1,7 @@
 export enum Game {
   SEVEN_DAYS_TO_DIE = "SEVEN_DAYS_TO_DIE",
   ARK_SURVIVAL_EVOLVED = "ARK_SURVIVAL_EVOLVED",
+  ARK_SURVIVAL_ASCENDED = "ARK_SURVIVAL_ASCENDED",
   DAYZ = "DAYZ",
   MINECRAFT = "MINECRAFT",
   PALWORLD = "PALWORLD",
